Series_of_Work/PC-Agent/requirements_mac.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.9.5
aiosignal==1.3.1
alibabacloud_credentials==0.3.6
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.3
alibabacloud-ocr-api20210707==3.1.2
alibabacloud_openapi_util==0.2.2
alibabacloud-tea==0.4.2
alibabacloud_tea_openapi==0.3.13
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
comtypes==1.4.10
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
dashscope==1.20.0
datasets==2.18.0
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
filelock==3.15.3
flatbuffers==1.12
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
gast==0.4.0
google-auth==2.30.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.23.4
idna==3.7
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==7.2.0
Jinja2==3.1.4
jiter==0.9.0
jmespath==0.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lmdb==1.6.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
modelscope==1.15.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
networkx==3.3
numpy==1.24.3
oauthlib==3.2.2
openai==1.66.2
opt-einsum==3.3.0
optree==0.11.0
oss2==2.18.6
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pip==25.0
platformdirs==4.2.2
propcache==0.3.0
protobuf==3.19.6
psutil==7.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
RapidFuzz==3.12.2
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
rubicon-objc==0.4.9
safetensors==0.4.3
scikit-image==0.25.2
scipy==1.13.1
sentencepiece==0.2.0
setuptools==75.8.0
shapely==2.0.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
supervision==0.21.0
sympy==1.12.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.9.0
termcolor==2.4.0
tf-keras==2.15.0
tf-slim==1.1.0
tifffile==2025.2.18
timm==1.0.7
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.45.1
wrapt==1.16.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
